requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
cachetools==5.5.1
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.7
cryptography==41.0.5
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.104.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.17.0
fsspec==2025.2.0
future==1.0.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-speech==2.30.0
googleapis-common-protos==1.67.0rc1
googletrans==4.0.0rc1
greenlet==3.0.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httptools==0.6.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
Mako==1.3.0
MarkupSafe==2.1.3
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
orjson==3.9.9
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pytest==7.4.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
textblob==0.19.0
tiktoken==0.8.0
tomli==2.0.1
tqdm==4.67.1
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.3.0
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
